Simon

+++ BookWormz.Services/CommentService.cs
@@ -30,9 +30,6 @@ namespace BookWormz.Services
 
             using (var ctx = new ApplicationDbContext())
             {
-                foreach (Comment comment in ctx.Comments)
-                    if (comment.Id == entity.ExchangeId)
-                        return false;
 
                 ctx.Comments.Add(entity);
                 return ctx.SaveChanges() == 1;
@@ -80,7 +77,7 @@ namespace BookWormz.Services
                     Id = entity.Id,
                     Text = entity.Text,
                     //Using ternary incase of nulled commenter
-                    CommentorsName = entity.Commenter != null ? entity.Commenter.FullName : "Unknown",
+                    CommentersName = entity.Commenter != null ? entity.Commenter.FullName : "Unknown",
                     Replies = AddReplies(entity.Replies)
                 };
                 return detailedComment;
@@ -94,6 +91,10 @@ namespace BookWormz.Services
             {
                 var entity = ctx.Comments.Single(e => e.Id == id);
 
+                //Make sure only commenter can update comment
+                if (entity.CommenterId != _userId)
+                    return false;
+
                 entity.Text = comment.CommentText;
 
                 return ctx.SaveChanges() == 1;
@@ -106,6 +107,11 @@ namespace BookWormz.Services
             using (var ctx = new ApplicationDbContext())
             {
                 var entity = ctx.Comments.Single(e => e.Id == Id);
+
+                //make sure only commenter can delete comment
+                if (entity.CommenterId != _userId)
+                    return false;
+
                 ctx.Comments.Remove(entity);
 
                 return ctx.SaveChanges() == 1;
@@ -124,7 +130,7 @@ namespace BookWormz.Services
             foreach(var reply in replies)
             {
                 var DetailedReply = new ReplyDetail { Id = reply.Id, Text = reply.Text,
-                    //Using ternary incase of comment not having author(corrupt data)
+                    //Using ternary incase of comment not having author
                     CommentorsName = (reply.Comment != null ? reply.Commenter.FullName : "Unknown") };
                 DetailedReply.Replies = AddReplies(reply.Replies);
                 DetailedReplies.Add(DetailedReply);
